+++ config/webpack-components.js
@@ -32,31 +32,31 @@ module.exports = {
 
    module: {
       rules: [{
-            test: /\.ts$/,
-            use: [{
-                  loader: 'awesome-typescript-loader?declaration=false',
-                  options: {
-                     configFileName: 'tsconfig.lib.json'
-                  }
-               },
-               {
-                  loader: 'angular2-template-loader'
-               }
-            ],
-            exclude: [/\.(spec|e2e)\.ts$/]
+         test: /\.ts$/,
+         use: [{
+            loader: 'awesome-typescript-loader?declaration=false',
+            options: {
+               configFileName: 'tsconfig.lib.json'
+            }
          },
          {
-            test: /\.html$/,
-            use: ['raw-loader']
-         },
-         {
-            test: /\.css$/,
-            use: ['to-string-loader', 'css-loader']
-         },
-         {
-            test: /\.scss$/,
-            use: ['to-string-loader', 'css-loader', 'sass-loader']
+            loader: 'angular2-template-loader'
          }
+         ],
+         exclude: [/\.(spec|e2e)\.ts$/]
+      },
+      {
+         test: /\.html$/,
+         use: ['raw-loader']
+      },
+      {
+         test: /\.css$/,
+         use: ['to-string-loader', 'css-loader']
+      },
+      {
+         test: /\.scss$/,
+         use: ['to-string-loader', 'css-loader', 'sass-loader']
+      }
       ]
    },
 
@@ -115,28 +115,6 @@ module.exports = {
             },
 
          }
-      }),
-      new UglifyJsPlugin({
-         beautify: false, //prod
-         output: {
-            comments: false
-         }, //prod
-         mangle: {
-            screw_ie8: true
-         }, //prod
-         compress: {
-            screw_ie8: true,
-            warnings: false,
-            conditionals: true,
-            unused: true,
-            comparisons: true,
-            sequences: true,
-            dead_code: true,
-            evaluate: true,
-            if_return: true,
-            join_vars: true,
-            negate_iife: false
-         },
-      }),
+      })
    ]
 };