+++ src/st-radio/index.ts
@@ -1,0 +1,1 @@
+export { StRadioModule } from './st-radio.module';