+++ src/st-radio/st-radio.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { StRadioComponent } from './st-radio.component';
+import { StRadioGroupComponent } from './st-radio-group.component';
+
+
+@NgModule({
+   imports: [
+      CommonModule
+   ],
+   exports: [
+      StRadioComponent, StRadioGroupComponent
+   ],
+   declarations: [
+      StRadioComponent, StRadioGroupComponent
+   ],
+   providers: []
+})
+export class StRadioModule { }