+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export { StRegEx } from './st-regex';
 export * from './event-window-manager';
 export * from './egeo-resolver';
 export { EgeoUtils } from './egeo-utils';
+export * from './unique-dispatcher';